backend/requirements.txt
fastapi>=0.115,<1
uvicorn>=0.32,<1
sqlalchemy>=2.0,<3
apscheduler>=3.10,<4
python-dotenv>=1.0,<2
